COMPLETED BACKEND

+++ backend/controllers/product.controller.js
@@ -49,10 +49,14 @@ export const updateProducts = async(req,res)=>{
 export const deleteProducts = async(req,res)=>{
     const {id} = req.params
 
+    if (!mongoose.Types.ObjectId.isValid(id)) { // check if the id is valid
+        return res.status(404).json({success:false,message:"Invalid product id"})
+    }
+
     try {
         await Product.findByIdAndDelete(id)
         res.status(200).json({success:true,message:"product deleted"})
     } catch (error) {
-        res.status(404).json({success:false,message:"product not found"})
+        res.status(500).json({success:false,message:"server error"})
     }
 }
